feat: add columns while edit table

+++ frontend/src/TooljetDatabase/PageBody/TableListItem/index.jsx
@@ -1,26 +1,31 @@
 import React, { useState, useContext } from 'react';
 import cx from 'classnames';
+
 import { toast } from 'react-hot-toast';
 import { tooljetDatabaseService } from '@/_services';
 import { ListItemPopover } from './ActionsPopover';
-import Drawer from '@/_ui/Drawer';
-import EditTableForm from '../../Forms/CreateTableForm';
 import { TooljetDatabaseContext } from '../../index';
 
+import Drawer from '@/_ui/Drawer';
+import EditTableForm from '../../Forms/TableForm';
+
 export const ListItem = ({ active, onClick, text = '', onDeleteCallback }) => {
-  const { organizationId, setTables } = useContext(TooljetDatabaseContext);
+  const { organizationId, columns, selectedTable, setTables } = useContext(TooljetDatabaseContext);
   const [isEditTableDrawerOpen, setIsEditTableDrawerOpen] = useState(false);
 
-  const handleDelete = async () => {
-    const { error } = await tooljetDatabaseService.deleteTable(organizationId, text);
+  const handleDeleteTable = async () => {
+    const shouldDelete = confirm(`Are you sure you want to delete the table "${text}"?`);
+    if (shouldDelete) {
+      const { error } = await tooljetDatabaseService.deleteTable(organizationId, text);
 
-    if (error) {
-      toast.error(error?.message ?? `Failed to delete table "${text}"`);
-      return;
+      if (error) {
+        toast.error(error?.message ?? `Failed to delete table "${text}"`);
+        return;
+      }
+
+      toast.success(`${text} deleted successfully`);
+      onDeleteCallback && onDeleteCallback();
     }
-
-    toast.success(`${text} deleted successfully`);
-    onDeleteCallback && onDeleteCallback();
   };
 
   const handleEdit = async (tableName) => {
@@ -32,6 +37,12 @@ export const ListItem = ({ active, onClick, text = '', onDeleteCallback }) => {
     // toast.success(`Edited table "${selectedTable}"`);
   };
 
+  const formColumns = columns.reduce((acc, column, currentIndex) => {
+    // todo: add data_type from table metadata api
+    acc[currentIndex] = { column_name: column.Header, data_type: 'todo' };
+    return acc;
+  }, {});
+
   return (
     <div
       className={cx('list-group-item cursor-pointer list-group-item-action text-capitalize', { active })}
@@ -39,10 +50,12 @@ export const ListItem = ({ active, onClick, text = '', onDeleteCallback }) => {
     >
       {text}
       <div className="float-right cursor-pointer">
-        <ListItemPopover onEdit={() => setIsEditTableDrawerOpen(true)} onDelete={handleDelete} />
+        <ListItemPopover onEdit={() => setIsEditTableDrawerOpen(true)} onDelete={handleDeleteTable} />
       </div>
       <Drawer isOpen={isEditTableDrawerOpen} onClose={() => setIsEditTableDrawerOpen(false)} position="right">
         <EditTableForm
+          selectedColumns={formColumns}
+          selectedTable={selectedTable}
           onCreate={() => {
             tooljetDatabaseService.findAll(organizationId).then(({ data = [] }) => {
               if (Array.isArray(data?.result) && data.result.length > 0) {
